css/css-fonts/math-script-level-and-math-style/discrete-no-interpolation.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/4441">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
<script>
test_no_interpolation({
  property: 'math-shift',
  from: 'initial',
  to: 'compact'
});

test_no_interpolation({
  property: 'math-style',
  from: 'initial',
  to: 'compact'
});
</script>
